requirements/default.txt
numpy==1.15.3
matplotlib==2.0.2
dendropy==4.3.0
pandas==0.23.4
h5py==2.8.0
bokeh==0.13.0
seaborn==0.8.1
biom-format==2.1.7
scikit-bio==0.5.1
